backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation & Serialization
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# File handling
python-magic==0.4.27
Pillow==11.1.0

# Communication
httpx==0.28.1
jinja2==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
